Restructure timing, cleanup
