+++ src/common/readout/vmm3/VMM3Parser.cpp
@@ -14,6 +14,8 @@
 
 namespace ESSReadout {
 
+  using namespace esstime;
+
 // #undef TRC_LEVEL
 // #define TRC_LEVEL TRC_L_DEB
