+++ src/common/readout/vmm3/test/VMM3ParserTest.cpp
@@ -7,7 +7,6 @@
 ///
 //===----------------------------------------------------------------------===//
 
-#include "ess/ESSTime.h"
 #include <common/readout/ess/Parser.h>
 #include <common/readout/vmm3/VMM3Parser.h>
 #include <common/readout/vmm3/test/VMM3ParserTestData.h>
