+++ src/common/time/CMakeLists.txt
@@ -1,0 +1,27 @@
+
+#=============================================================================
+# Add ess timin modules
+#=============================================================================
+
+include_directories(.)
+
+set(esstime_obj_SRC
+  ESSTime.cpp
+  Timer.cpp
+  TimeString.cpp
+  TSCTimer.cpp
+)
+
+set(esstime_obj_INC
+ESSTime.h
+Timer.h
+TimeString.h
+TSCTimer.h
+)
+
+add_library(EssTimingLib OBJECT
+  ${esstime_obj_SRC}
+  ${esstime_obj_INC}
+  )
+
+enable_coverage_flags(EssTimingLib)
