+++ src/modules/caen/CaenInstrument.h
@@ -16,7 +16,6 @@
 #include <caen/CaenCounters.h>
 #include <caen/geometry/Config.h>
 #include <caen/readout/Readout.h>
-#include <common/readout/ess/ESSTime.h>
 #include <common/readout/ess/Parser.h>
 #include <cspec/geometry/CspecGeometry.h>
 #include <loki/geometry/LokiGeometry.h>
