+++ src/modules/dream/DreamInstrument.h
@@ -11,7 +11,6 @@
 
 #pragma once
 
-#include <common/readout/ess/ESSTime.h>
 #include <common/readout/ess/Parser.h>
 #include <dream/Counters.h>
 #include <dream/DreamBase.h> // to get DreamSettings
