+++ src/modules/nmx/generators/MultiHitReadoutGenerator.h
@@ -10,7 +10,6 @@
 
 #pragma once
 
-#include "common/readout/ess/ESSTime.h"
 #include <common/readout/vmm3/VMM3Parser.h>
 #include <common/testutils/DataFuzzer.h>
 #include <generators/essudpgen/ReadoutGeneratorBase.h>
