+++ src/modules/nmx/generators/SmileReadoutGenerator.h
@@ -10,13 +10,12 @@
 
 #pragma once
 
-#include "common/readout/ess/ESSTime.h"
 #include <common/readout/vmm3/VMM3Parser.h>
 #include <common/testutils/DataFuzzer.h>
 #include <generators/essudpgen/ReadoutGeneratorBase.h>
 
 namespace Nmx {
-  
+
 class SmileReadoutGenerator : public ReadoutGeneratorBase {
 public:
   using ReadoutGeneratorBase::ReadoutGeneratorBase;
