+++ src/modules/nmx/generators/TrackReadoutGenerator.cpp
@@ -1,4 +1,4 @@
-// Copyright (C) 2022 - 2023 European Spallation Source, ERIC. See LICENSE file
+// Copyright (C) 2022 - 2024 European Spallation Source, ERIC. See LICENSE file
 //===----------------------------------------------------------------------===//
 ///
 /// \file
@@ -10,15 +10,7 @@
 // GCOVR_EXCL_START
 
 #include <common/debug/Trace.h>
-#include <math.h>
 #include <modules/nmx/generators/TrackReadoutGenerator.h>
-#include <time.h>
-
-#include <cassert>
-#include <cstdint>
-#include <cstdio>
-#include <cstring>
-#include <stdexcept>
 
 // #undef TRC_LEVEL
 // #define TRC_LEVEL TRC_L_DEB
