+++ src/modules/trex/TREXInstrument.cpp
@@ -8,7 +8,6 @@
 ///
 //===----------------------------------------------------------------------===//
 
-#include "common/time/TimeNano.h"
 #include <assert.h>
 #include <common/debug/Log.h>
 #include <common/debug/Trace.h>
