+++ src/modules/trex/TREXInstrument.h
@@ -12,7 +12,6 @@
 
 #include <common/kafka/EV44Serializer.h>
 #include <common/monitor/Histogram.h>
-#include <common/readout/ess/ESSTime.h>
 #include <common/readout/ess/Parser.h>
 #include <common/readout/vmm3/Hybrid.h>
 #include <common/readout/vmm3/Readout.h>
