+++ src/modules/trex/generators/LETReadoutGenerator.h
@@ -10,7 +10,6 @@
 
 #pragma once
 
-#include "common/readout/ess/ESSTime.h"
 #include <common/readout/vmm3/VMM3Parser.h>
 #include <common/testutils/DataFuzzer.h>
 #include <generators/essudpgen/ReadoutGeneratorBase.h>
@@ -23,7 +22,6 @@ public:
 
 protected:
   void generateData() override;
-  ESSReadout::ESSTime::PulseTime generatePulseTime() override;
 
   uint64_t GlobalReadout{0};
 };
